pyproject.toml
[project]
name = "trading-bots"
version = "0.1.0"
description = "A project containing various deployable and runtime configurable trading strategies"
authors = [
    {name = "Moses Wuniche A.",email = "alhassanmoses.am@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-binance (>=1.0.28,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "backtrader (>=1.9.78.123,<2.0.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
types-requests = "^2.32.0.20240622"
types-pyyaml = "^6.0.12.20250402"
black = "^25.1.0"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-binance = "^1.0.19"
pandas = "^2.2.2"
numpy = "^1.26.4"
boto3 = "^1.34.98"
pyyaml = "^6.0.1"
matplotlib = "^3.9.0"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
